backend/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ello-bot"
dynamic = ["version"]
description = "ello-bot backend service"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "苍镜月" }]
requires-python = ">=3.12"

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.129.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
]

[project.urls]
Homepage = "https://github.com/LunaY77/ello-bot"
Documentation = "https://github.com/LunaY77/ello-bot"
Repository = "https://github.com/LunaY77/ello-bot"

[project.scripts]
ello-server = "app.main:main"
lint = "app.tools.scripts:lint"

[tool.setuptools.dynamic]
version = { attr = "app.__version__" }

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = [
    "E501", # 行长度交给格式化或按需放开
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
